requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
apscheduler
pydantic-settings
cuid
chromadb
fastmcp
google-generativeai
